mlx5: fold dump helpers into dumps
Move the safe-access helpers into their only caller and remove the
unnecessary compatibility module.

Orabug: 39751588
